mysite/requirements.txt
django-crispy-forms==1.6.0
django-formtools==1.0
django-password-reset==0.8.2
django-jquery==1.12.2
django-highcharts==0.1.7
django-smart_selects==1.5.2
geopy
reportlab
django-ipware

